+++ apps/entity-service/src/app.module.ts
@@ -39,9 +39,6 @@ import { AggregateReportsService } from "./entities/aggregate-reports.service";
 import { ReportingFrameworksController } from "./reportingFrameworks/reporting-frameworks.controller";
 import { ReportingFrameworksService } from "./reportingFrameworks/reporting-frameworks.service";
 import { ExportImageService } from "./entities/export-image.service";
-import { EntityCsvExportService } from "./entities/entity-csv-export.service";
-import { FinancialReportController } from "./entities/financial-report.controller";
-import { SrpReportController } from "./entities/srp-report.controller";
 
 @Module({
   imports: [
@@ -64,8 +61,6 @@ import { SrpReportController } from "./entities/srp-report.controller";
     RemindersController,
     AuditStatusController, // must be before EntitiesController
     AggregateReportsController, // must be before EntitiesController
-    FinancialReportController, // must be before EntitiesController
-    SrpReportController, // must be before EntitiesController
     EntitiesController,
     FormDataController, // must be before entity association controller.
     UpdateRequestsController, // must be before entity association controller.
@@ -93,8 +88,7 @@ import { SrpReportController } from "./entities/srp-report.controller";
     FormDataService,
     EntitiesQueueProcessor,
     ReportingFrameworksService,
-    ExportImageService,
-    EntityCsvExportService
+    ExportImageService
   ]
 })
 export class AppModule {}
