+++ libs/common/src/lib/export/csv-export.service.ts
@@ -4,31 +4,44 @@ import { stringify } from "csv-stringify";
 import { FileService } from "../file/file.service";
 import { ConfigService } from "@nestjs/config";
 import { FileDownloadDto } from "../dto/file-download.dto";
-import { Dictionary, pick } from "lodash";
+import { Dictionary, groupBy, pick } from "lodash";
 import { Model } from "sequelize";
 import { DateTime } from "luxon";
 import { Response } from "express";
-
-function serializeCell(value: unknown): string | number {
-  if (value == null) return "";
-  if (value instanceof Date) return DateTime.fromJSDate(value).toISODate() ?? "";
-  if (Array.isArray(value)) {
-    return value.map(v => (v == null ? "" : serializeCell(v))).join("; ");
-  }
-  if (typeof value === "object") {
-    return JSON.stringify(value);
-  }
-  return value as string | number;
-}
+import {
+  getExportHeading,
+  getLinkedFieldConfig,
+  getModelAttribute,
+  LinkedFieldSpecification,
+  ModelAttribute
+} from "../linkedFields";
+import { Form, FormQuestion, FormSection, Media } from "@terramatch-microservices/database/entities";
+import { FormModels, LinkedAnswerCollector } from "../linkedFields/linkedAnswerCollector";
+import { FrameworkKey } from "@terramatch-microservices/database/constants";
+import { MediaService } from "../media/media.service";
+import { isField, isFile } from "@terramatch-microservices/database/constants/linked-fields";
+import { FormModelType } from "@terramatch-microservices/database/constants/entities";
+import { isNotNull } from "@terramatch-microservices/database/types/array";
 
 type StreamWriter = {
   addRow: (model: Model, additional?: Dictionary<unknown>) => void;
   close: () => void;
 };
 
+type FormQuestionExportMapping = {
+  questionUuid: string;
+  heading: string;
+  attribute?: ModelAttribute;
+  config?: LinkedFieldSpecification;
+};
+
 @Injectable()
 export class CsvExportService {
-  constructor(private readonly fileService: FileService, private readonly configService: ConfigService) {}
+  constructor(
+    private readonly fileService: FileService,
+    private readonly configService: ConfigService,
+    private readonly mediaService: MediaService
+  ) {}
 
   get bucket() {
     const bucket = this.configService.get<string>("AWS_BUCKET");
@@ -60,24 +73,61 @@ export class CsvExportService {
     return this.createStreamWriter(response, columns);
   }
 
-  private createStreamWriter<T extends NodeJS.WritableStream>(
-    destination: T,
-    columns: Dictionary<string>
-  ): StreamWriter {
-    const stringifier = stringify({ header: true, columns });
-    stringifier.pipe(destination);
+  async getFormQuestionsForExport(form: Form) {
+    const sections = await FormSection.findAll({ where: { formId: form.uuid }, order: [["order", "ASC"]] });
+    const questions = await FormQuestion.forForm(form.uuid).findAll({ order: [["order", "ASC"]] });
+    const sectionQuestions = groupBy(
+      questions.filter(({ parentId }) => parentId == null),
+      "formSectionId"
+    );
+    const childQuestions = groupBy(
+      questions.filter(({ parentId }) => parentId != null),
+      "parentId"
+    );
 
-    const keys = Object.keys(columns);
-    return {
-      addRow: (model: Model, additional?: Dictionary<unknown>) => {
-        const row = Object.entries({ ...pick(model, keys), ...additional }).reduce(
-          (acc, [key, value]) => ({ ...acc, [key]: serializeCell(value) }),
-          {}
-        );
-        stringifier.write(row);
-      },
-      close: () => stringifier.end()
-    };
+    const mappings: FormQuestionExportMapping[] = [];
+
+    for (const section of sections) {
+      for (const question of sectionQuestions[`${section.id}`] ?? []) {
+        this.addQuestionToMapping(mappings, question);
+
+        for (const child of childQuestions[`${question.id}`] ?? []) {
+          this.addQuestionToMapping(mappings, child);
+        }
+      }
+    }
+
+    return mappings;
+  }
+
+  getAttributes(mappings: FormQuestionExportMapping[], model: FormModelType) {
+    return mappings
+      .filter(({ attribute }) => attribute?.model === model)
+      .map(({ attribute }) => attribute?.attribute)
+      .filter(isNotNull);
+  }
+
+  async collectFormCells(mappings: FormQuestionExportMapping[], models: FormModels, frameworkKey?: FrameworkKey) {
+    const collector = new LinkedAnswerCollector(this.mediaService);
+    for (const mapping of mappings) {
+      if (mapping.config == null) continue;
+
+      const { model, field } = mapping.config;
+      if (isField(field)) collector.fields.addField(field, model, mapping.questionUuid);
+      else if (isFile(field)) collector.files.addField(field, model, mapping.questionUuid);
+      else collector[field.resource].addField(field, model, mapping.questionUuid);
+    }
+
+    const answers: Dictionary<unknown> = {};
+    await collector.collect(answers, models, { forExport: true, frameworkKey });
+
+    // Some collectors need the original question UUID to correctly map their readable export data,
+    // so we have to pass the question UUID into the collector and then re-map to our headings after
+    // data collection.
+    return Object.entries(answers).reduce((acc, [questionUuid, value]) => {
+      const { heading } = mappings.find(mapping => questionUuid === mapping.questionUuid) ?? {};
+      return heading == null ? acc : { ...acc, [heading]: value };
+    }, {});
   }
 
   /**
@@ -89,10 +139,55 @@ export class CsvExportService {
     const filteredRows = rows.map(row => {
       const filteredRow: Record<string, string | number> = {};
       for (const { key } of columnsArray) {
-        filteredRow[key] = serializeCell(row[key]);
+        filteredRow[key] = this.serializeCell(row[key]);
       }
       return filteredRow;
     });
     return stringifySync(filteredRows, { header: true, columns: columnsArray });
+  }
+
+  private createStreamWriter<T extends NodeJS.WritableStream>(
+    destination: T,
+    columns: Dictionary<string>
+  ): StreamWriter {
+    const stringifier = stringify({ header: true, columns });
+    stringifier.pipe(destination);
+
+    const keys = Object.keys(columns);
+    return {
+      addRow: (model: Model, additional?: Dictionary<unknown>) => {
+        const row = Object.entries({ ...pick(model, keys), ...additional }).reduce(
+          (acc, [key, value]) => ({ ...acc, [key]: this.serializeCell(value) }),
+          {}
+        );
+        stringifier.write(row);
+      },
+      close: () => stringifier.end()
+    };
+  }
+
+  private serializeCell(value: unknown): string | number {
+    if (value == null) return "";
+    if (value instanceof Date) return DateTime.fromJSDate(value).toISODate() ?? "";
+    if (value instanceof Media) return this.mediaService.getUrl(value) ?? "";
+    if (Array.isArray(value)) {
+      return value.map(v => (v == null ? "" : this.serializeCell(v))).join("|");
+    }
+    if (typeof value === "object") {
+      return JSON.stringify(value);
+    }
+    return value as string | number;
+  }
+
+  private addQuestionToMapping(mappings: FormQuestionExportMapping[], question: FormQuestion) {
+    if (question.linkedFieldKey == null || question.inputType === "tableInput") return;
+
+    const config = getLinkedFieldConfig(question.linkedFieldKey);
+    mappings.push({
+      questionUuid: question.uuid,
+      heading: getExportHeading(config),
+      attribute: getModelAttribute(config),
+      config
+    });
   }
 }
