+++ libs/common/src/lib/policies/financial-report.policy.ts
@@ -5,13 +5,10 @@ import { AWAITING_APPROVAL, DUE, STARTED } from "@terramatch-microservices/datab
 export class FinancialReportPolicy extends UserPermissionsPolicy {
   async addRules() {
     if (this.frameworks.length > 0) {
-      this.builder.can(
-        ["read", "delete", "update", "approve", "updateAnswers", "sendReminder", "export"],
-        FinancialReport,
-        {
-          frameworkKey: { $in: this.frameworks }
-        }
-      );
+      this.builder.can(["exportAll"], FinancialReport);
+      this.builder.can(["read", "delete", "update", "approve", "updateAnswers", "sendReminder"], FinancialReport, {
+        frameworkKey: { $in: this.frameworks }
+      });
     }
 
     if (this.permissions.includes("manage-own")) {
